Fix script value semantics, return parsing, and LSP symbol analysis
Preserve JSON null in loops, normalize numeric truthiness, and concatenate JSON strings consistently. Report function arity errors as usage errors and accept bare return before block endings.

Resolve document-local functions, analyze nested commands, and preserve case-sensitive variable references and hover ranges. Add value-origin matrix coverage and script regression tests, and document behavior and remaining limitations.

Validation: 2679 offline tests passed, 2 interactive console tests skipped; warning-free builds and clean git diff --check.

+++ CosmosDBShell/Azure.Data.Cosmos.Shell.Parser/ShellObject/ShellJson.cs
@@ -66,7 +66,7 @@ internal class ShellJson : ShellObject
 
                 if (this.Value.ValueKind == JsonValueKind.Number)
                 {
-                    return this.Value.GetInt32() != 0;
+                    return this.Value.GetDouble() != 0;
                 }
 
                 throw new InvalidOperationException($"Cannot convert JSON {this.Value.ValueKind} to boolean");
